tree-construction/domjs.dat

#data
<svg><![CDATA[foo
bar]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo
bar"

#data
<svg><![CDATA[foobar]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo
bar"

#data
<svg><![CDATA[foo
bar]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo
bar"
